+++ packages/shared/src/catalog/index.ts
@@ -18,6 +18,7 @@ export * from "./primitives/link";
 export * from "./primitives/listBox";
 export * from "./primitives/menu";
 export * from "./primitives/numberField";
+export * from "./primitives/popover";
 export * from "./primitives/radio";
 export * from "./primitives/radioGroup";
 export * from "./primitives/searchField";